refactor: refactor cuda_kernel/neuron_kernel
