+++ spikingjelly/activation_based/cuda_kernel/__init__.py
@@ -1,2 +1,21 @@
 from . import auto_cuda  # noqa
-from .neuron_kernel import *
+from .neuron_kernel import (
+    multistep_eif_ptt,
+    multistep_if_ptt,
+    multistep_izhikevich_ptt,
+    multistep_lif_ptt,
+    multistep_plif_ptt,
+    multistep_qif_ptt,
+    save_cuda_codes,
+)
+
+__all__ = [
+    "auto_cuda",
+    "save_cuda_codes",
+    "multistep_if_ptt",
+    "multistep_lif_ptt",
+    "multistep_plif_ptt",
+    "multistep_qif_ptt",
+    "multistep_izhikevich_ptt",
+    "multistep_eif_ptt",
+]
